Separate backend routes to private and public folders
